requirements.txt
Flask==2.0.2
requests==2.26.0
pyjwt==2.3.0
cryptography==3.4.7
python-dotenv==0.19.1
